Fix pre-commits

+++ src/modules/training/datasets/main_dataset.py
@@ -28,7 +28,7 @@ class MainDataset(Dataset):  # type: ignore[type-arg]
     steps: list[TrainingBlock] | None = None
 
     X: XData | None = None
-    y: npt.NDArray[np.float32] | None = None
+    y: npt.NDArray[np.int8] | None = None
     indices: npt.NDArray[np.int32] | None = None
 
     def __post_init__(self) -> None:
@@ -44,7 +44,7 @@ class MainDataset(Dataset):  # type: ignore[type-arg]
         # Setup Pipeline
         self.setup_pipeline(use_augmentations=False)
 
-    def initialize(self, X: XData, y: npt.NDArray[np.float32] | None = None, indices: list[int] | npt.NDArray[np.int32] | None = None) -> None:
+    def initialize(self, X: XData, y: npt.NDArray[np.int8] | None = None, indices: list[int] | npt.NDArray[np.int32] | None = None) -> None:
         """Set up the dataset for training."""
         self.X = X
         self.y = y
